Added new group detail component
